I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary
for now, merge into's parallelism depends on the source query, we can improve this.
```sql
-- before improve
mysql> insert overwrite 
    ->     test_merge_into_target 
    -> select 
    ->     id, max(v1), max(v2) 
    -> from 
    ->     (
    ->         select 
    ->             floor(rand() * 80000) as id, 
    ->             floor(rand() * 1000) as v1, 
    ->             floor(rand() * 1000) as v2 
    ->         from 
    ->             numbers(40000)) 
    -> group by 
    ->     id
    -> ;
Query OK, 31407 rows affected (0.33 sec)

mysql> insert overwrite 
    ->     test_merge_into_source 
    -> select 
    ->     floor(rand() * 80000) as id, 
    ->     floor(rand() * 1000) as v1, 
    ->     floor(rand() * 1000) as v2 
    -> from 
    ->     numbers(40000)
    -> ;
Query OK, 40000 rows affected (0.22 sec)

mysql> MERGE INTO 
    ->     test_merge_into_target AS T
    -> USING 
    ->     (
    ->         SELECT 
    ->             id, max(v1) as v1, max(v2) as v2 
    ->         FROM 
    ->             test_merge_into_source 
    ->         GROUP BY 
    ->             id
    ->     ) AS S
    -> ON 
    ->     T.id = S.id
    -> WHEN MATCHED THEN
    ->     UPDATE SET
    ->         T.id = S.id, T.v1 = S.v1, T.v2 = S.v2 
    -> WHEN NOT MATCHED THEN
    ->     INSERT 
    ->         (id, v1, v2) 
    ->     VALUES 
    ->         (S.id, S.v1, S.v2);
+-------------------------+------------------------+
| number of rows inserted | number of rows updated |
+-------------------------+------------------------+
|                   19020 |                  12411 |
+-------------------------+------------------------+
1 row in set (7.30 sec)
Read 71407 rows, 1.66 MiB in 7.260 sec., 9.84 thousand rows/sec., 234.14 KiB/sec.
```

```sql
-- after improve
mysql> insert overwrite 
    ->     test_merge_into_target 
    -> select 
    ->     id, max(v1), max(v2) 
    -> from 
    ->     (
    ->         select 
    ->             floor(rand() * 80000) as id, 
    ->             floor(rand() * 1000) as v1, 
    ->             floor(rand() * 1000) as v2 
    ->         from 
    ->             numbers(40000)) 
    -> group by 
    ->     id
    -> ;
Query OK, 31516 rows affected (0.45 sec)

mysql> insert overwrite 
    ->     test_merge_into_source 
    -> select 
    ->     floor(rand() * 80000) as id, 
    ->     floor(rand() * 1000) as v1, 
    ->     floor(rand() * 1000) as v2 
    -> from 
    ->     numbers(40000)
    -> ;
Query OK, 40000 rows affected (0.22 sec)

mysql> MERGE INTO 
    ->     test_merge_into_target AS T
    -> USING 
    ->     (
    ->         SELECT 
    ->             id, max(v1) as v1, max(v2) as v2 
    ->         FROM 
    ->             test_merge_into_source 
    ->         GROUP BY 
    ->             id
    ->     ) AS S
    -> ON 
    ->     T.id = S.id
    -> WHEN MATCHED THEN
    ->     UPDATE SET
    ->         T.id = S.id, T.v1 = S.v1, T.v2 = S.v2 
    -> WHEN NOT MATCHED THEN
    ->     INSERT 
    ->         (id, v1, v2) 
    ->     VALUES 
    ->         (S.id, S.v1, S.v2);
+-------------------------+------------------------+
| number of rows inserted | number of rows updated |
+-------------------------+------------------------+
|                   19070 |                  12481 |
+-------------------------+------------------------+
1 row in set (3.89 sec)
Read 71516 rows, 1.66 MiB in 3.850 sec., 18.58 thousand rows/sec., 442.18 KiB/sec.
```

## Tests

- [ ] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [ ] No Test  - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/14574)
<!-- Reviewable:end -->